Convo4-AI-Chat/requirements.txt
openai
langchain
python-dotenv
langchain-community
